+++ src/backend/NES/PAK/Mappers/M002.hpp
@@ -31,7 +31,7 @@ namespace NES_NS {
              * @param cBnk Number of CHR-ROM banks
              * @param cMem Reference to CHR memory on gamepak.
              */
-            M002(u8 pBnk, vector<u8>& pMem, u8 cBnk, vector<u8>& cMem) :
+            M002(u16 pBnk, vector<u8>& pMem, u16 cBnk, vector<u8>& cMem) :
                 Mapper(pBnk, pMem, cBnk, cMem) {};
 
             /**
